Removed Enemy and Player interaction + fixed health and oxygen items
There were issues with the prefab, i cant put script in the prefab so that the animations will work and the code. Apparently unity cant support it? <https://forum.unity.com/threads/cant-add-script-component-to-prefab.141831/>  I have to put the prefab in the scene and put the code to the prefab and use that prefab to the enemy spawner, but with that, as I said, if we "kill" one of them they will all be dead or they "kill" us and instead of -5 health its 5*15(i limited it to 15 spawns) so basically you;re dead if you go near one of them